Refactor presentation routes
