+++ backend/routers/__init__.py
@@ -1,10 +1,16 @@
 from .presentations import router as presentations_router
+from .presentation_steps import router as presentation_steps_router
+from .presentation_modify import router as presentation_modify_router
+from .presentation_images import router as presentation_images_router
 from .images import router as images_router
 from .logos import router as logos_router
 from .pptx import router as pptx_router
 
 __all__ = [
     "presentations_router",
+    "presentation_steps_router",
+    "presentation_modify_router",
+    "presentation_images_router",
     "images_router",
     "logos_router",
     "pptx_router",
